Extract component to ContractDropdown.tsx
